lights_and_shadows

+++ modules/engine/component/spriteRenderer.cpp
@@ -1,5 +1,7 @@
 #include "spriteRenderer.h"
 
+#include <type_traits>
+
 #include "data/primitives/planePrimitiveData.h"
 #include "rendering/renderer.h"
 #include "rendering/geometry/primitiveGenerator.h"
@@ -8,12 +10,73 @@ namespace BreadEngine {
     DEFINE_STATIC_PROPS(SpriteRenderer)
     REGISTER_COMPONENT(SpriteRenderer)
 
+    static_assert(std::is_nothrow_move_constructible_v<SpriteRenderer>);
+    static_assert(std::is_nothrow_move_assignable_v<SpriteRenderer>);
+
     /// Pixels per world unit for a sprite quad.
     constexpr float SPRITE_SCALE = .01f;
 
     SpriteRenderer::SpriteRenderer(Node *owner)
     {
         _owner = owner;
+    }
+
+    SpriteRenderer::SpriteRenderer(const SpriteRenderer &other)
+    {
+        Component::operator=(other);
+        _material = other._material;
+        _textureAsset = other._textureAsset;
+    }
+
+    SpriteRenderer &SpriteRenderer::operator=(const SpriteRenderer &other)
+    {
+        if (this == &other) return *this;
+
+        unload();
+        Component::operator=(other);
+        _material = other._material;
+        _textureAsset = other._textureAsset;
+        return *this;
+    }
+
+    SpriteRenderer::SpriteRenderer(SpriteRenderer &&other) noexcept
+    {
+        Component::operator=(other);
+        _material = std::move(other._material);
+        _mesh = other._mesh;
+        _textureAsset = other._textureAsset;
+        _textureMaterial = other._textureMaterial;
+        _acquiredTexture = other._acquiredTexture;
+        _quadTexture = other._quadTexture;
+        _isLoaded = other._isLoaded;
+
+        other._mesh = {};
+        other._textureMaterial = {};
+        other._acquiredTexture = nullptr;
+        other._quadTexture = {};
+        other._isLoaded = false;
+    }
+
+    SpriteRenderer &SpriteRenderer::operator=(SpriteRenderer &&other) noexcept
+    {
+        if (this == &other) return *this;
+
+        unload();
+        Component::operator=(other);
+        _material = std::move(other._material);
+        _mesh = other._mesh;
+        _textureAsset = other._textureAsset;
+        _textureMaterial = other._textureMaterial;
+        _acquiredTexture = other._acquiredTexture;
+        _quadTexture = other._quadTexture;
+        _isLoaded = other._isLoaded;
+
+        other._mesh = {};
+        other._textureMaterial = {};
+        other._acquiredTexture = nullptr;
+        other._quadTexture = {};
+        other._isLoaded = false;
+        return *this;
     }
 
     void SpriteRenderer::onDestroy()
@@ -46,10 +109,13 @@ namespace BreadEngine {
     {
         if (!_isLoaded) return;
 
-        auto &renderer = Renderer::get();
-        renderer.destroyMesh(_mesh);
-        renderer.destroyMaterial(_textureMaterial);
-        if (_acquiredTexture != nullptr) _acquiredTexture->release();
+        if (Renderer::isAlive())
+        {
+            auto &renderer = Renderer::get();
+            renderer.destroyMesh(_mesh);
+            renderer.destroyMaterial(_textureMaterial);
+            if (_acquiredTexture != nullptr) _acquiredTexture->release();
+        }
 
         _mesh = {};
         _textureMaterial = {};
